Refactor code to better handle math overflows
Replaced the previous "expect" calls with a match construct to handle number overflows, which provides safer and more elegant error handling. Added a new MathError enumeration to classify typical math-related errors, improving the clarity of error messages. Changed the name of "num.rs" into "u256_math.rs" to represent its functionality more accurately.
