+++ utils/src/lib.rs
@@ -2,8 +2,9 @@
 
 pub mod bump;
 pub mod constant;
-pub mod num;
+pub mod math_errors;
 pub mod storage;
 pub mod storage_errors;
 pub mod test_utils;
 pub mod token_utils;
+pub mod u256_math;
