+++ utils/src/math_errors.rs
@@ -1,0 +1,8 @@
+use soroban_sdk::contracterror;
+
+#[contracterror]
+#[derive(Copy, Clone, Debug, Eq, PartialEq, PartialOrd, Ord)]
+#[repr(u32)]
+pub enum MathError {
+    NumberOverflow = 510,
+}
